pyproject.toml
[project]
name = "steramlit-langchain-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "chunking-evaluation",
    "ipykernel>=6.29.5",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.25",
    "openai>=1.90.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]

[tool.uv.sources]
chunking-evaluation = { git = "https://github.com/brandonstarxel/chunking_evaluation.git" }
